Step4 edit and finish
